refactor, remove useless and add license header
