+++ packages/ruling/tests/data/eslint-to-sonar-id.json
